@@ -1,3 +1,6 @@
 export const environment = {
-  production: true
+  production: false,
+  BASE_URL: 'http://178.128.217.187:5010/api/',
+  SERVER_URL : 'http://localhost:5000/'
+
 };
