@@ -4,7 +4,7 @@
 
 export const environment = {
   production: false,
-  BASE_URL: 'http://localhost:5010/api/',
+  BASE_URL: 'http://178.128.217.187:5010/api/',
   SERVER_URL : 'http://localhost:5000/'
 };
 
